requirements.txt
# requirements.txt
fastapi==0.97.0
pydantic==1.10.9
psycopg2-binary==2.9.6
python-dotenv==1.0.0
mangum==0.17.0
uvicorn==0.22.0

